# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  devops-cd_internal_dto.BuildSummary:
    properties:
      build_created:
        type: string
      build_number:
        type: integer
      build_status:
        type: string
      commit_author:
        type: string
      commit_message:
        type: string
      commit_sha:
        type: string
      id:
        type: integer
      image_tag:
        type: string
    type: object
...
  devops-cd_internal_dto.CreateProjectRequest:
    properties:
      allowed_env_clusters:
        additionalProperties:
          items:
            type: string
          type: array
        description: '允许的环境集群配置: {"pre": ["cluster-a"], "prod": ["cluster-b"]}'
        type: object
      create_default_team:
        type: boolean
      default_env_clusters:
        additionalProperties:
          items:
            type: string
          type: array
        description: 项目默认环境集群配置(必须是 allowed_env_clusters 的子集)
        type: object
      description:
        type: string
      name:
        maxLength: 100
        type: string
      owner_name:
        maxLength: 100
        type: string
    required:
    - name
    type: object
...
  devops-cd_internal_dto.GetReleaseAppRequest:
    properties:
      id:
        description: 发布应用ID
        type: integer
    required:
    - id
    type: object
...
  devops-cd_internal_dto.PageResponse:
    properties:
      items: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  devops-cd_internal_dto.ProjectAvailableEnvClustersResponse:
    properties:
      allowed_env_clusters:
        additionalProperties:
          items:
            type: string
          type: array
        description: '完整配置: {"pre": ["cluster-a"], "prod": ["cluster-b"]}'
        type: object
      available_clusters:
        description: 如果指定了env,返回该环境下可用集群列表
        items:
          type: string
        type: array
    type: object
  devops-cd_internal_dto.ProjectEnvConfigRequest:
    properties:
      allow_clusters:
        items:
          type: string
        type: array
      chart_repo_url:
        maxLength: 255
        type: string
      default_clusters:
        items:
          type: string
        type: array
      deployment_name_template:
        maxLength: 255
        type: string
      namespace:
        maxLength: 63
        type: string
      values_path_template:
        maxLength: 255
        type: string
      values_repo_url:
        maxLength: 255
        type: string
    required:
    - allow_clusters
    type: object
  devops-cd_internal_dto.ProjectEnvConfigResponse:
    properties:
      allow_clusters:
        items:
          type: string
        type: array
      chart_repo_url:
        type: string
      created_at:
        type: string
      default_clusters:
        items:
          type: string
        type: array
      deployment_name_template:
        type: string
      env:
        type: string
      id:
        type: integer
      namespace:
        type: string
      project_id:
        type: integer
      updated_at:
        type: string
      values_path_template:
        type: string
      values_repo_url:
        type: string
    type: object
  devops-cd_internal_dto.ProjectResponse:
    properties:
      allowed_env_clusters:
        additionalProperties:
          items:
            type: string
          type: array
        description: 允许的环境集群配置
        type: object
      created_at:
        type: string
      default_env_clusters:
        additionalProperties:
          items:
            type: string
          type: array
        description: 项目默认环境集群配置(必须是 allowed_env_clusters 的子集)
        type: object
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      owner_name:
        type: string
      teams:
        items:
          $ref: '#/definitions/devops-cd_internal_dto.TeamResponse'
        type: array
      updated_at:
        type: string
    type: object
...
  devops-cd_internal_dto.ReleaseAppResponse:
    properties:
      app_id:
        type: integer
      app_name:
        description: 应用信息
        type: string
      app_project:
        type: string
      app_status:
        type: integer
      app_type:
        type: string
      batch_id:
        type: integer
      build_id:
        description: 关联的构建ID
        type: integer
      build_number:
        description: 构建信息（通过 build 关联获取，可选）
        type: integer
      build_status:
        description: 构建状态
        type: string
      build_time:
        type: string
      commit_branch:
        description: 分支
        type: string
      commit_message:
        description: commit 信息
        type: string
      commit_sha:
        description: commit SHA
        type: string
      created_at:
        description: 时间信息
        type: string
      default_depends_on:
        description: 依赖信息
        items:
          type: integer
        type: array
      deployed_tag:
        description: 应用当前部署的镜像标签（从 applications 表获取）
        type: string
      id:
        description: ReleaseApp 基本信息
        type: integer
      image_url:
        description: 完整镜像地址
        type: string
      is_locked:
        description: 是否已锁定（封板后为true）
        type: boolean
      latest_build_id:
        description: 版本信息
        type: integer
      previous_deployed_tag:
        description: 部署前的版本（封板时记录）
        type: string
      reason:
        type: string
      recent_builds:
        description: 最近的构建记录（自上次部署以来，最多15条）
        items:
          $ref: '#/definitions/devops-cd_internal_dto.BuildSummary'
        type: array
      release_notes:
        description: 发布信息
        type: string
      repo_full_name:
        type: string
      repo_id:
        description: 仓库信息
        type: integer
      repo_name:
        type: string
      skip_pre_env:
        description: 是否跳过预发布环境（封板时从app_env_configs计算得出）
        type: boolean
      status:
        type: integer
      target_tag:
        description: 目标部署版本（封板时固定，部署期间代表期望版本，部署完成后代表已部署版本）
        type: string
      team_id:
        description: 团队信息
        type: integer
      team_name:
        type: string
      temp_depends_on:
        items:
          type: integer
        type: array
      updated_at:
        type: string
    type: object
  devops-cd_internal_dto.ReleaseDependenciesResponse:
    properties:
      app_id:
        type: integer
      batch_id:
        type: integer
      default_depends_on:
        items:
          type: integer
        type: array
      release_app_id:
        type: integer
      temp_depends_on:
        items:
          type: integer
        type: array
      updated_at:
        type: string
    type: object
...
  devops-cd_internal_dto.SwitchVersionRequest:
    properties:
      batch_id:
        description: 批次ID
        type: integer
      build_id:
        description: 目标build id
        type: integer
      operator:
        description: 操作人
        type: string
      reason:
        description: 触发原因（可选）
        type: string
      release_app_id:
        description: 发布应用ID
        type: integer
    required:
    - batch_id
    - build_id
    - operator
    - release_app_id
    type: object
  devops-cd_internal_dto.TeamResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      leader_name:
        type: string
      name:
        type: string
      project_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  devops-cd_internal_dto.UpdateBuildsRequest:
    properties:
      batch_id:
        type: integer
      build_changes:
        additionalProperties:
          format: int64
          type: integer
        description: 'key: app_id, value: build_id'
        type: object
      operator:
        type: string
    required:
    - batch_id
    - build_changes
    - operator
    type: object
  devops-cd_internal_dto.UpdateProjectEnvConfigsRequest:
    properties:
      configs:
        additionalProperties:
          $ref: '#/definitions/devops-cd_internal_dto.ProjectEnvConfigRequest'
        description: 'key: env (pre/prod)'
        type: object
    required:
    - configs
    type: object
  devops-cd_internal_dto.UpdateProjectRequest:
    properties:
      allowed_env_clusters:
        additionalProperties:
          items:
            type: string
          type: array
        description: 允许的环境集群配置
        type: object
      default_env_clusters:
        additionalProperties:
          items:
            type: string
          type: array
        description: 项目默认环境集群配置(必须是 allowed_env_clusters 的子集)
        type: object
      description:
        type: string
      id:
        type: integer
      name:
        maxLength: 100
        type: string
      owner_name:
        maxLength: 100
        type: string
    required:
    - id
    type: object
  devops-cd_internal_dto.UpdateReleaseDependenciesRequest:
    properties:
      batch_id:
        type: integer
      operator:
        type: string
      temp_depends_on:
        items:
          type: integer
        type: array
    required:
    - batch_id
    - operator
    type: object
...
  devops-cd_pkg_utils.Response:
    properties:
      code:
        type: integer
      data: {}
      detail:
        description: 详细错误信息（可选）
        type: string
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: |-
    DevOps 持续交付平台 API 文档
    提供代码库管理、应用管理、批次管理、部署管理等功能
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: DevOps CD API
  version: "1.0"
paths:
...
  /api/v1/batch/release_app:
    put:
      consumes:
      - application/json
      description: 批量更新批次中应用的构建版本等信息（仅草稿状态可修改）
      parameters:
      - description: 更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/devops-cd_internal_dto.UpdateBuildsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 批次状态不允许修改
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 更新失败
          schema:
            additionalProperties: true
            type: object
      summary: 更新批次发布应用
      tags:
      - 批次管理
...
  /api/v1/project:
    get:
      consumes:
      - application/json
      parameters:
      - description: 项目ID
        format: int64
        in: query
        name: id
        required: true
        type: integer
      - description: 是否包含团队列表
        in: query
        name: with_teams
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/devops-cd_pkg_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/devops-cd_internal_dto.ProjectResponse'
              type: object
      summary: 获取项目详情
      tags:
      - Project
  /api/v1/project/{id}/env:
    get:
      consumes:
      - application/json
      parameters:
      - description: 项目ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/devops-cd_pkg_utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/devops-cd_internal_dto.ProjectEnvConfigResponse'
                  type: array
              type: object
      summary: 获取项目的环境配置
      tags:
      - Project
    put:
      consumes:
      - application/json
      parameters:
      - description: 项目ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: 环境配置
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/devops-cd_internal_dto.UpdateProjectEnvConfigsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/devops-cd_pkg_utils.Response'
      summary: 批量更新项目的环境配置
      tags:
      - Project
  /api/v1/projects:
    get:
      consumes:
      - application/json
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 关键字搜索
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/devops-cd_pkg_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/devops-cd_internal_dto.PageResponse'
              type: object
      summary: 获取项目列表（无分页参数时返回所有项目，有分页参数时返回分页数据）
      tags:
      - Project
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建项目请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/devops-cd_internal_dto.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/devops-cd_pkg_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/devops-cd_internal_dto.ProjectResponse'
              type: object
      summary: 创建项目
      tags:
      - Project
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新项目请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/devops-cd_internal_dto.UpdateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/devops-cd_pkg_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/devops-cd_internal_dto.ProjectResponse'
              type: object
      summary: 更新项目
      tags:
      - Project
  /api/v1/projects/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 项目ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/devops-cd_pkg_utils.Response'
      summary: 删除项目
      tags:
      - Project
  /api/v1/projects/available-env-clusters:
    get:
      consumes:
      - application/json
      parameters:
      - description: 项目ID
        format: int64
        in: query
        name: project_id
        required: true
        type: integer
      - description: 环境名称(pre/prod),不传返回全部配置
        in: query
        name: env
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/devops-cd_pkg_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/devops-cd_internal_dto.ProjectAvailableEnvClustersResponse'
              type: object
      summary: 获取项目可用的环境集群配置
      tags:
      - Project
  /api/v1/release_app:
    get:
      consumes:
      - application/json
      parameters:
      - description: 发布应用ID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/devops-cd_internal_dto.GetReleaseAppRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/devops-cd_pkg_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/devops-cd_internal_dto.ReleaseAppResponse'
              type: object
      summary: 获取发布应用详情
      tags:
      - ReleaseApp
  /api/v1/release_app/{id}/dependencies:
    put:
      consumes:
      - application/json
      parameters:
      - description: ReleaseApp ID
        in: path
        name: id
        required: true
        type: integer
      - description: 依赖更新请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/devops-cd_internal_dto.UpdateReleaseDependenciesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/devops-cd_pkg_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/devops-cd_internal_dto.ReleaseDependenciesResponse'
              type: object
      summary: 更新发布应用临时依赖
      tags:
      - ReleaseApp
  /api/v1/release_app/trigger_deploy:
    post:
      consumes:
      - application/json
      parameters:
      - description: 切换请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/devops-cd_internal_dto.SwitchVersionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/devops-cd_pkg_utils.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 切换版本
      tags:
      - ReleaseApp
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"