# internal/api/handler/project_handler.go
package handler

import (
	"devops-cd/pkg/responses"
	"net/http"
	"strconv"

	"github.com/gin-gonic/gin"

	"devops-cd/internal/dto"
	"devops-cd/internal/service"
)

type ProjectHandler struct {
	projectService service.ProjectService
}

func NewProjectHandler(projectService service.ProjectService) *ProjectHandler {
	return &ProjectHandler{
		projectService: projectService,
	}
}

// Create 创建项目
// @Summary 创建项目
// @Tags Project
// @Accept json
// @Produce json
// @Param request body dto.CreateProjectRequest true "创建项目请求"
// @Success 200 {object} utils.Response{data=dto.ProjectResponse}
// @Router /api/v1/projects [post]
func (h *ProjectHandler) Create(c *gin.Context) {
	var req dto.CreateProjectRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		responses.ErrorWithDetail(c, http.StatusBadRequest, "请求参数错误", err.Error())
		return
	}

	project, err := h.projectService.Create(&req)
	if err != nil {
		responses.Error(c, err)
		return
	}

	responses.Success(c, project)
}

// GetByID 获取项目详情
// @Summary 获取项目详情
// @Tags Project
// @Accept json
// @Produce json
// @Param id query int64 true "项目ID"
// @Param with_teams query bool false "是否包含团队列表"
// @Success 200 {object} utils.Response{data=dto.ProjectResponse}
// @Router /api/v1/project [get]
func (h *ProjectHandler) GetByID(c *gin.Context) {
	var req dto.GetProjectRequest
	if err := c.ShouldBindQuery(&req); err != nil {
		responses.ErrorWithDetail(c, http.StatusBadRequest, "请求参数错误", err.Error())
		return
	}

	project, err := h.projectService.GetByID(req.ID, req.WithTeams)
	if err != nil {
		responses.Error(c, err)
		return
	}

	responses.Success(c, project)
}

// List 获取项目列表
// @Summary 获取项目列表（无分页参数时返回所有项目，有分页参数时返回分页数据）
// @Tags Project
// @Accept json
// @Produce json
// @Param page query int false "页码"
// @Param page_size query int false "每页数量"
// @Param keyword query string false "关键字搜索"
// @Success 200 {object} utils.Response{data=[]dto.ProjectSimpleResponse}
// @Success 200 {object} utils.Response{data=dto.PageResponse}
// @Router /api/v1/projects [get]
func (h *ProjectHandler) List(c *gin.Context) {
	var query dto.ProjectListQuery
	if err := c.ShouldBindQuery(&query); err != nil {
		responses.ErrorWithDetail(c, http.StatusBadRequest, "请求参数错误", err.Error())
		return
	}

	// 如果没有分页参数，返回所有项目简化列表（用于下拉选择）
	if query.Page == 0 && query.PageSize == 0 {
		projects, err := h.projectService.ListAll()
		if err != nil {
			responses.Error(c, err)
			return
		}
		responses.Success(c, projects)
		return
	}

	// 有分页参数，返回分页数据
	projects, total, err := h.projectService.List(&query)
	if err != nil {
		responses.Error(c, err)
		return
	}

	responses.Success(c, dto.NewPageResponse(projects, total, query.GetPage(), query.GetPageSize()))
}

// Update 更新项目
// @Summary 更新项目
// @Tags Project
// @Accept json
// @Produce json
// @Param request body dto.UpdateProjectRequest true "更新项目请求"
// @Success 200 {object} utils.Response{data=dto.ProjectResponse}
// @Router /api/v1/projects [put]
func (h *ProjectHandler) Update(c *gin.Context) {
	var req dto.UpdateProjectRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		responses.ErrorWithDetail(c, http.StatusBadRequest, "请求参数错误", err.Error())
		return
	}

	project, err := h.projectService.Update(req.ID, &req)
	if err != nil {
		responses.Error(c, err)
		return
	}

	responses.Success(c, project)
}

// Delete 删除项目
// @Summary 删除项目
// @Tags Project
// @Accept json
// @Produce json
// @Param id path int64 true "项目ID"
// @Success 200 {object} utils.Response
// @Router /api/v1/projects/{id} [delete]
func (h *ProjectHandler) Delete(c *gin.Context) {
	idStr := c.Param("id")
	id, err := strconv.ParseInt(idStr, 10, 64)
	if err != nil {
		responses.ErrorWithDetail(c, http.StatusBadRequest, "无效的项目ID", err.Error())
		return
	}

	if err := h.projectService.Delete(id); err != nil {
		responses.Error(c, err)
		return
	}

	responses.Success(c, nil)
}

// GetAvailableEnvClusters 获取项目可用的环境集群配置
// @Summary 获取项目可用的环境集群配置
// @Tags Project
// @Accept json
// @Produce json
// @Param project_id query int64 true "项目ID"
// @Param env query string false "环境名称(pre/prod),不传返回全部配置"
// @Success 200 {object} utils.Response{data=dto.ProjectAvailableEnvClustersResponse}
// @Router /api/v1/projects/available-env-clusters [get]
func (h *ProjectHandler) GetAvailableEnvClusters(c *gin.Context) {
	var req dto.GetProjectAvailableEnvClustersRequest
	if err := c.ShouldBindQuery(&req); err != nil {
		responses.ErrorWithDetail(c, http.StatusBadRequest, "请求参数错误", err.Error())
		return
	}

	resp, err := h.projectService.GetAvailableEnvClusters(req.ProjectID, req.Env)
	if err != nil {
		responses.Error(c, err)
		return
	}

	responses.Success(c, resp)
}

// GetEnvConfigs 获取项目的环境配置
// @Summary 获取项目的环境配置
// @Tags Project
// @Accept json
// @Produce json
// @Param id path int64 true "项目ID"
// @Success 200 {object} utils.Response{data=[]dto.ProjectEnvConfigResponse}
// @Router /api/v1/project/{id}/env [get]
func (h *ProjectHandler) GetEnvConfigs(c *gin.Context) {
	idStr := c.Param("id")
	id, err := strconv.ParseInt(idStr, 10, 64)
	if err != nil {
		responses.ErrorWithDetail(c, http.StatusBadRequest, "无效的项目ID", err.Error())
		return
	}

	configs, err := h.projectService.GetEnvConfigs(id)
	if err != nil {
		responses.Error(c, err)
		return
	}

	responses.Success(c, configs)
}

// UpdateEnvConfigs 批量更新项目的环境配置
// @Summary 批量更新项目的环境配置
// @Tags Project
// @Accept json
// @Produce json
// @Param id path int64 true "项目ID"
// @Param request body dto.UpdateProjectEnvConfigsRequest true "环境配置"
// @Success 200 {object} utils.Response
// @Router /api/v1/project/{id}/env [put]
func (h *ProjectHandler) UpdateEnvConfigs(c *gin.Context) {
	idStr := c.Param("id")
	id, err := strconv.ParseInt(idStr, 10, 64)
	if err != nil {
		responses.ErrorWithDetail(c, http.StatusBadRequest, "无效的项目ID", err.Error())
		return
	}

	var req dto.UpdateProjectEnvConfigsRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		responses.ErrorWithDetail(c, http.StatusBadRequest, "请求参数错误", err.Error())
		return
	}

	if err := h.projectService.UpdateEnvConfigs(id, req.Configs); err != nil {
		responses.Error(c, err)
		return
	}

	responses.Success(c, nil)
}